Move CheckTargetMutations into pipeline_utils and setup v2
